List of the board members
I copied from Elin's email telling the result of the election, sent on the 11th March.
Not sure if we need to spell out all the acronyms of the affiliations (or even different affiliation, f. eks. Lex's should be MN rather than IBV, in the sense of consistency of the level of the organisation if the others are like USIT, UB, and UV?). Also, should we have a link to each employee page?